Resolved review comments.
