+++ src/licit/extensions/tableEx/tableEx.ts
@@ -11,32 +11,7 @@ import  { Table,
 } from '@tiptap/extension-table';
 import {Node as ProseMirrorNode} from 'prosemirror-model';
 import {TextSelection} from 'prosemirror-state';
-
-const NUMERIC_VALUE_PATTERN = /^-?\d+(\.\d+)?$/;
-
-const normalizeValue = (value: unknown): string | null => {
-  if (value === null || value === undefined) {
-    return null;
-  }
-
-  if (typeof value !== 'string' && typeof value !== 'number') {
-    return null;
-  }
-
-  const normalized = `${value}`.trim();
-  return normalized.length ? normalized : null;
-};
-
-const normalizeCssSize = (value: unknown): string | null => {
-  const normalized = normalizeValue(value);
-  if (!normalized) {
-    return null;
-  }
-
-  return NUMERIC_VALUE_PATTERN.test(normalized)
-    ? `${normalized}px`
-    : normalized;
-};
+import {normalizeCssSize, normalizeValue} from '../table.utils';
 
 class TableViewEx extends TiptapTableView {
   constructor(node: ProseMirrorNode, cellMinWidth: number) {
