+++ src/licit/ui/Tabledetails.css
@@ -14,7 +14,6 @@
 }
 
 .czi-row--inset {
-  background: #1e1e1e;
   border: 1px solid #555;
   border-radius: 4px;
   padding: 6px 10px;
